Code cleanup & PM's working better
